000-099-investing/004-stocks-tracker/requirements.txt
# Shared core module (install from 000-shared-core)
# pip install -e ../000-shared-core

# Core dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
requests>=2.28.0

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Retry logic for API calls
tenacity==8.2.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# Development tools
ruff==0.1.6
